fix(apps/client): HASHI-81 인기맛집 코드리뷰 반영
