+++ apps/client/src/features/restaurantList/components/RestaurantCard.tsx
@@ -1,7 +1,7 @@
 import { StarFillIcon } from '@hashi/hds-icons'
 
-import type { Restaurant } from './types'
-import { RestaurantImageList } from './RestaurantImageList'
+import { RestaurantImageList } from '@/features/restaurantList/components/RestaurantImageList'
+import type { Restaurant } from '@/features/restaurantList/types'
 
 type RestaurantCardProps = {
   restaurant: Restaurant
@@ -17,26 +17,26 @@ export const RestaurantCard = ({
   }
 
   return (
-    <li className="border-warm-gray-50 h-[300px] w-full border-b">
+    <li className="border-warm-gray-50 w-full border-b py-4.75">
       <button
-        className="flex h-full w-full flex-col pt-[19px] text-left"
+        className="flex w-full flex-col text-left"
         onClick={handleClickRestaurant}
         type="button"
       >
-        <span className="typo-body-3 text-cool-gray-900 line-clamp-1 h-[19px]">
+        <span className="typo-body-3 text-cool-gray-900 line-clamp-1">
           {restaurant.name}
         </span>
         <span className="text-primary-200 mt-1 flex h-5 items-center">
           <StarFillIcon
             aria-hidden="true"
-            className="text-primary-400 size-[18px] shrink-0"
+            className="text-primary-400 size-4.5 shrink-0"
           />
           <span className="typo-body-3 ml-px">{restaurant.rating}</span>
-          <span className="typo-body-7 ml-[5px]">
+          <span className="typo-body-7 ml-1.25">
             {restaurant.region} · {restaurant.category}
           </span>
         </span>
-        <span className="mt-2.5 w-full">
+        <span className="mt-2.75 w-full">
           <RestaurantImageList
             images={restaurant.images}
             restaurantName={restaurant.name}
@@ -45,7 +45,7 @@ export const RestaurantCard = ({
         <span className="typo-body-7 text-primary-200 mt-3 line-clamp-2 w-full">
           {restaurant.description}
         </span>
-        <span className="mt-0.5 flex flex-wrap gap-x-2 gap-y-1">
+        <span className="mt-0.5 flex flex-wrap gap-2">
           {restaurant.hashtags.map((hashtag, index) => (
             <span
               className="typo-body-7 text-cool-gray-400"
