+++ apps/client/src/features/restaurantList/constants/index.ts
@@ -4,5 +4,4 @@ export {
   RESTAURANT_IMAGE_SLOT_COUNT,
   RESTAURANT_LIST_PAGE_SIZE,
 } from './layout'
-export { MOCK_RESTAURANTS } from './mockRestaurants'
 export { POPULAR_RESTAURANTS_SORT_OPTIONS } from './popularRestaurants'
