+++ apps/client/src/pages/hashiPick/HashiPick.spec.md
@@ -43,8 +43,9 @@
 - [x] `초기화`는 현재 열린 sheet의 draft 값만 기본값으로 되돌립니다.
 - [x] 필터 BottomSheet는 X 버튼으로만 닫힙니다.
 - [x] RestaurantCard는 반복 렌더링되고 카드 클릭 시 식당 상세 route로 이동합니다.
-- [x] 식당 이미지는 가로 스크롤 리스트로 표시합니다.
+- [x] 식당 이미지는 가로 스크롤 리스트로 표시하고 이미지가 없으면 공통 `DefaultImage` fallback을 사용합니다.
 - [x] 식당 리스트는 mock data 25개를 10개 단위로 무한스크롤 렌더링합니다.
+- [x] sticky header는 `z-fixed` 토큰을 사용합니다.
 
 ## Data Dependencies
 
@@ -74,16 +75,21 @@
 - server state: none
 - derived state:
   - category filter label: `전체`일 때 `음식 장르 선택`, 그 외 selected label
+- owner:
+  - list state and handlers are owned by `useRestaurantListPage`
 
 ## UI Structure
 
 ```text
 HashiPickPage
-  Header
-  RestaurantFilterBar
-  RestaurantCard list
-  FilterBottomSheet(sort)
-  FilterBottomSheet(category)
+  RestaurantListPage
+    Header
+    RestaurantFilterBar
+    RestaurantCard list
+      RestaurantImageList
+        DefaultImage fallback
+    FilterBottomSheet(sort)
+    FilterBottomSheet(category)
 ```
 
 ## Component Mapping
@@ -96,13 +102,18 @@ HashiPickPage
 - app shared component:
   - `FilterBottomSheet`
 - feature component:
+  - `RestaurantListPage`
   - `RestaurantFilterBar`
   - `RestaurantCard`
+  - `RestaurantImageList`
 - feature hook:
+  - `useRestaurantListPage`
   - `useInfiniteRestaurantList`
+- feature mock:
+  - `mocks/restaurantList.mock.ts`
 - page-local component: none
 - icon:
-  - `BackIcon`
+- `BackIcon`
   - `TapDownIcon`
   - `CheckIcon`
   - `StarFillIcon`
@@ -120,7 +131,7 @@ HashiPickPage
 
 ## Verification
 
-- [x] `corepack pnpm --filter @hashi/client test -- HashiPickPage.test.tsx FilterBottomSheet.test.tsx`
+- [x] `corepack pnpm --filter @hashi/client test -- HashiPickPage.test.tsx PopularRestaurantsPage.test.tsx FilterBottomSheet.test.tsx`
 - [x] `corepack pnpm --filter @hashi/client lint`
 - [x] `corepack pnpm --filter @hashi/client typecheck`
 - [x] `corepack pnpm --filter @hashi/client build`
