+++ apps/client/src/shared/components/filterBottomSheet/FilterBottomSheet.tsx
@@ -32,7 +32,7 @@ export const FilterBottomSheet = ({
   onApply,
 }: FilterBottomSheetProps) => {
   const sheetFooter = (
-    <div className="grid grid-cols-2 gap-[13px]">
+    <div className="grid grid-cols-2 gap-3.25">
       <Button onClick={onReset} size="md" variant="neutral" width="full">
         초기화
       </Button>
@@ -43,50 +43,48 @@ export const FilterBottomSheet = ({
   )
   return (
     <BottomSheet
-      closeOnEscape={false}
-      closeOnOverlayClick={false}
       aria-label={title}
       className={cn(
         'flex max-h-[calc(100dvh-40px)] flex-col rounded-t-[10px]',
         maxHeightClassName,
       )}
-      contentClassName="flex-1 overflow-y-auto"
       footer={sheetFooter}
       open={open}
-      overlayClassName="bg-[#1d1d1d]/70"
       onOpenChange={onOpenChange}
       title={<span className="typo-sub-header-2">{title}</span>}
     >
-      <ul
-        className="flex flex-col gap-[5px] pt-[40px] pb-1"
+      <div
+        className="max-h-[calc(100dvh-180px)] overflow-y-auto"
         data-testid="filter-bottom-sheet-content"
       >
-        {options.map((option) => {
-          const isSelected = option.value === selectedValue
+        <ul className="flex flex-col gap-1.25 pt-10 pb-1">
+          {options.map((option) => {
+            const isSelected = option.value === selectedValue
 
-          return (
-            <li key={option.value}>
-              <button
-                aria-pressed={isSelected}
-                className={cn(
-                  'flex min-h-5 w-full items-center justify-between py-[10px] text-black',
-                  isSelected ? 'typo-body-3' : 'typo-body-4',
-                )}
-                onClick={() => onSelect(option.value)}
-                type="button"
-              >
-                <span>{option.label}</span>
-                {isSelected && (
-                  <CheckIcon
-                    aria-hidden="true"
-                    className="text-cool-gray-700 size-5 shrink-0"
-                  />
-                )}
-              </button>
-            </li>
-          )
-        })}
-      </ul>
+            return (
+              <li key={option.value}>
+                <button
+                  aria-pressed={isSelected}
+                  className={cn(
+                    'flex min-h-5 w-full items-center justify-between py-2.5 text-black',
+                    isSelected ? 'typo-body-3' : 'typo-body-4',
+                  )}
+                  onClick={() => onSelect(option.value)}
+                  type="button"
+                >
+                  <span>{option.label}</span>
+                  {isSelected && (
+                    <CheckIcon
+                      aria-hidden="true"
+                      className="text-cool-gray-700 size-5 shrink-0"
+                    />
+                  )}
+                </button>
+              </li>
+            )
+          })}
+        </ul>
+      </div>
     </BottomSheet>
   )
 }
